requirements.txt
scipy>=1.5.5
numpy>=1.18.3
tensorflow_gpu==2.8.0
matplotlib>=3.2.1
